Move source selection helpers to sfs.util
